refactor(minimax-m3): name the MSA kernel modules by pipeline stage
The kernels package holds forward and backward modules side by side, and
two names did not say which side they serve. msa_patch.py only patches
the upstream forward package, and msa_grad_finalize_sm100.py is the last
launch of the backward, the counterpart of
msa_backward_preprocess_sm100.py. Rename them to msa_forward_patch.py
and msa_backward_postprocess_sm100.py so the directory reads as the
pipeline it is: forward patch, backward preprocess, backward kernel,
backward postprocess.

Both module bodies are byte-identical to their previous paths. Only the
import lines in msa_backward_sm100.py and _msa.py, the backward kernel's
module docstring and the README follow the new names.
