+++ nemo_automodel/components/models/minimax_m3_vl/kernels/msa_forward_patch.py
